pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aivy-tutor"
version = "0.1.0"
description = "Aivy Tutor with memory management system"
requires-python = ">=3.9,<4.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
    "numpy",
    "pandas",
    "google-generativeai>=0.3.0",
    "jina",
    "mem0ai>=0.1.37",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.9.1",
    "pymilvus>=2.4.3",
    "pydantic>=2.7.3",
    "langchain-community>=0.3.1",
    "fastapi",
    "uvicorn"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff"
]

[tool.ruff]
line-length = 120
select = ["E", "F"]
ignore = []

[tool.black]
line-length = 120
target-version = ["py39"]